Skip proctor authorization at end if use_grade_proctor is "No"
In that case that user_grade_proctor is "No" instead of changing the
assignment type from 'proctored_gateway' to 'gateway' the first time
that proctor authorization has been obtained, skip obtaining proctor
authorization when "Grade Test" is hit.  This means that if the user
leaves the quiz for any reason the user will still need proctor
authorization again to re-enter the quiz.  However, when the user hits
"Grade Test" the test will just be graded without the need of proctor
authorization.