Fix rendering static images when checking answers in the PG problem editor.
This fixes issue #2556 essentially as @somiaj suggests (although there
is no need to first check that `$rh->{fileName}` is defined).

The reason that `$rh->{fileName}` is defined is because the
`default.html.ep` template always adds the `fileName` hidden field even
when the `fileName` parameter is not defined in the original request. So
if you click "Check Answers" it will be defined in `RenderProblem.pm` on
that request.