Fix a rounding error with SingleProblemGrader point values.
When computing the point value in the SingleProblemGrader, the
database saves 6 decimal digits in the status field, while the
single problem grader only saved 2. This difference in rounding can
cause the computation of the point value with a stepSize of 0.25
to get get different values between the computed score and the
score saved in the database, causing the "Use points from last check"
button to show up and present a different rounded value.

This fixes the issue by saving the current score in a `$rawCurrnetScore`
variable that isn't rounded and using that for the point computation.
Note, `$rawCurrentScore` can have more than the 6 decimal digits in the
saved database score, but rounding this to match the database won't
improve the accuracy of the point value rounded to the nearest stepSize.