Change LTI update to allow selecting multiple sets.
Update the LTI Mass Update page to allow updating either all users
or a single user for multiple sets with a single update. To be able
to determine if all sets for a specific user are selected or not,
refactor the database calls to get the list of all users and their
assigned sets as part of the initialize method.