Allow viewing ProblemSet page for any valid set.
Rework the ProblemSet page so users can always see the set header
and additional set information for any set assigned to them.
Instead of giving an error message when a user tries to access
a set that is not open, lti restricted, conditional release,
or ip restricted, checkSet will set a flag which is used to
determine what information a user can see.

If a set not visible and a user cannot see hidden sets, only a
warning message is shown. In all other cases the set header
and any date information, such as when the set opens, closes,
etc, is shown. If the set is restricted due to a conditional
release, ip restrictions, or lti restrictions, a warning message
is shown informing the user of the restriction.

When a set is restricted, never show the set problems and always
show any previous taken set versions. This way users can still
access set versions they have taken for a restricted set from
a non restricted location.

Note, this is consistent with the permission `view_unoppend_sets`
description, which states it only configures if a set problems
can be seen. So this permission is no longer used to see if a
user can see set information such as open date and set header.

This also makes it so the right info panel div on the ProblemSet
page is only shown if the set header exists, and is not empty.

Translations were added to messages about IP restrictions.